Import functions from a module

Import specific functions from an outsider module to the Global Environment.

module_import(fname, repo)

Arguments

fname
Function name to import

repo
Module repo

Value

Function

Details

If program is successfully removed from your system, TRUE is returned else FALSE.

See also

Examples

library(outsider)
# simplest repo
repo <- 'dombennett/om..hello.world'
# install
module_install(repo = repo)

#> latest: Pulling from dombennett/om_hello.world
#> 124c757242f8: Already exists
#> 9d866f8bde2a: Already exists
#> fa3f2f277e67: Already exists
#> 398d32b153e8: Already exists
#> afde35469481: Already exists
#> 6376e5eb74c4: Pulling fs layer
#> 057c1f061444: Pulling fs layer
#> 057c1f061444: Verifying Checksum
#> 057c1f061444: Download complete
#> 6376e5eb74c4: Verifying Checksum
#> 6376e5eb74c4: Download complete
#> 6376e5eb74c4: Pull complete
#> 057c1f061444: Pull complete
#> Digest: sha256:f6770b2eea5e5f3d3934a2b1a14acf696675b2086d11227c27b096c9dfabb7dd
#> Status: Downloaded newer image for dombennett/om_hello.world:latest

#> [1] TRUE

# check if installed
(module_installed())

#> # A tibble: 1 x 2
#>   pkg                         repo                      
#>   <chr>                       <chr>                     
#> 1 om..hello.world..dombennett dombennett/om..hello.world

# get help
module_help(repo = repo)
# import function
hello_world <- module_import(fname = 'hello_world', repo = repo)
# run function
hello_world()

#> 1b53e5ebd5969d734ea2e20f4ff3cd49b09d31181a914e48f82fa50c993075c9
#> Hello world!
#> ------------
#> DISTRIB_ID=Ubuntu
#> DISTRIB_RELEASE=18.04
#> DISTRIB_CODENAME=bionic
#> DISTRIB_DESCRIPTION="Ubuntu 18.04.1 LTS"
#> hello.world_0
#> hello.world_0

# uninstall
rm(hello_world)
module_uninstall(repo = repo)

#> Untagged: dombennett/om_hello.world:latest
#> Untagged: dombennett/om_hello.world@sha256:f6770b2eea5e5f3d3934a2b1a14acf696675b2086d11227c27b096c9dfabb7dd
#> Deleted: sha256:c0cb087733b9adadb450360477d78b2a7e261f0bdec804f36ce3ba41902fc2d0
#> Deleted: sha256:42732d30db60051011756269795de113c9d8265adc4a5149f1f94bb7a07e4c13
#> Deleted: sha256:c80459014be1f9d2908724be50a24c545c98a9fb5a4c25a1effcec93ddcac6c8