Install an outsider module

Install a module through multiple different methods: via a code sharing site such as GitHub, a URL, a git repository or local filepath. The function will first install the R package and then build the Docker image. Docker image version is determined by "tag". To avoid pulling the image from DockerHub set "manual" to TRUE.

module_install(
  repo = NULL,
  url = NULL,
  filepath = NULL,
  git = NULL,
  service = c("github", "bitbucket", "gitlab"),
  tag = "latest",
  manual = FALSE,
  verbose = FALSE,
  force = FALSE,
  update = c("default", "ask", "always", "never")
)

Arguments

repo
Module repo, character.

url
URL to downloadable compressed (zip, tar or bzipped/gzipped) folder of a module, character.

filepath
Filepath to uncompressed directory of module, character.

git
URL to git repository

service
Code-sharing service. Character.

tag
Module version, default latest. Character.

manual
Build the docker image? Default FALSE. Logical.

verbose
Be verbose? Default FALSE.

force
Ignore warnings and install anyway? Default FALSE.

update
Update dependent R packages?

Value

Logical

Details

All installation options depend on the installation functions of remotes. E.g. GitHub packages are installed with install_github. See these functions for more details on the R package installation process.

Examples

library(outsider)
if (is_outsider_ready()) {
  # simplest repo
  repo <- 'dombennett/om..hello.world'
  # install
  module_install(repo = repo, force = TRUE, update = 'never')
  # is module_installed?
  (is_module_installed(repo = repo))
  # get help for package
  #module_help(repo = repo)
  # list functions available
  module_functions(repo = repo)
  # import
  hello_world <- module_import(fname = 'hello_world', repo = repo)
  # get help for function
  #module_help(repo = repo, fname = 'hello_world')
  # ?hello_world # also works
  # run function
  hello_world()

  # change verbosity settings
  # print nothing to console
  verbosity_set(show_program = FALSE, show_docker = FALSE)
  hello_world()
  # print everything to console
  verbosity_set(show_program = TRUE, show_docker = TRUE)
  hello_world()
  # write program output to a file
  verbosity_set(show_program = 'log.txt', show_docker = FALSE)
  hello_world()
  (readLines(con = 'log.txt'))
  file.remove('log.txt')

  # uninstall
  module_uninstall(repo)
}

#> cd88425c6b5d625d02d2245ba3589069a95da4496cf2bdc273e3d31deb0649c6
#> Hello world!
#> ------------
#> DISTRIB_ID=Ubuntu
#> DISTRIB_RELEASE=18.04
#> DISTRIB_CODENAME=bionic
#> DISTRIB_DESCRIPTION="Ubuntu 18.04.1 LTS"
#> om_hello_world_0
#> om_hello_world_0

#> Removing 'om..hello.world'

#> Removing package from ‘/private/var/folders/x9/m8kwpxps2v93xk52zqzm5lkh0000gp/T/RtmpWlb8ao/temp_libpath15a0651f2e52’
#> (as ‘lib’ is unspecified)