Test an outsider module

Ensure an outsider module builds, imports correctly and all its functions successfully complete.

.module_test(repo, verbose = FALSE)

Arguments

repo
Module repo

verbose
Print docker and program info to console

Value

Logical

Details

Success or fail, the module is uninstalled from the machine after the test is run.

See also